+++ main/ocr2_extract_text.cpp
@@ -106,7 +106,7 @@ int main(int ac, char **av)
 
     Image const img = image_from_file(av[3]);
     Bounds const bounds = img.bounds();
-    size_t x = 0;
+    unsigned x = 0;
 
     using resolution_clock = std::chrono::high_resolution_clock;
     auto t1 = resolution_clock::now();
